moved get_images_from_dir() to Resources.py
